% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyse_feedback_folder_1.R
\name{analyse_feedback_folder_1}
\alias{analyse_feedback_folder_1}
\title{Function to analyse data from SHiNeMaS using R packages \code{shinemas2R} and \code{PPBstats} regarding template_feedback_folder_1}
\usage{
analyse_feedback_folder_1(info_db, year = "2016")
}
\arguments{
\item{info_db}{used by shinemas2R::get.data() : a list with the following element to connect to the data base:
\itemize{
\item db_user     user name of SHiNeMaS
\item db_host     IP address of the computer where SHiNeMaS is. If local host db_host = "127.0.0.1"
\item db_name     name of the data base
\item db_password    your password to login. If no password is needed, put ""
}}

\item{year}{the year of the feedback folder}

\item{mc.cores}{The number of cores used for parallelisation of the computing}
}
\description{
Function to analyse data from SHiNeMaS using R packages \code{shinemas2R} and \code{PPBstats} regarding template_feedback_folder_1
}
\author{
Pierre Rivière, Gaëlle Van Frank
}

