% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feedback_folder_1.R
\name{feedback_folder_1}
\alias{feedback_folder_1}
\title{Function to get the "feedback file" containing the results of the year's trial for each farmer based on the analysis coming from \code{analyse_feedback_folder_1}}
\usage{
feedback_folder_1(dir = ".", person, out_analyse_feedback_folder_1)
}
\arguments{
\item{dir}{Directory where folder for each person is created}

\item{person}{The farmer's name}

\item{out_analyse_feedback_folder_1}{The outputs from \code{analyse_feedback_folder_1}}
}
\value{
Generate tex and pdf files
}
\description{
Function to get the "feedback file" containing the results of the year's trial for each farmer based on the analysis coming from \code{analyse_feedback_folder_1}
}
\details{
The function creates two folders :
\itemize{
 \item "tex_files" with the tex files used to create the pdf
 \item "feedback_folder" with, for each person, a folder with information coming from shinemas2R::get.pdf() see ?get.pdf for more details.
}
}
\author{
Pierre Rivière, Gaelle Van Frank
}

