% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_result_model.R
\name{get_result_model}
\alias{get_result_model}
\title{Select parts of models results from PPBstats::analyse.outputs (eigher parameters comparisons or MCMC).}
\usage{
get_result_model(res_model, data, type_result = "comp.mu", variable,
  param = NULL, year = NULL)
}
\arguments{
\item{res_model}{output from the \code{PPBstats::analyse.outputs} function}

\item{data}{output from the \code{shinemas2R::get.data} function}

\item{type_result}{the results wanted : parameters comparison ("comp.mu") or MCMC ("MCMC")}

\item{variable}{the variable}

\item{param}{if type_result is "MCMC", the parameter wanted ("mu","beta","sigma" if model 1 is used, "alpha","beta","theta")}

\item{year}{the year wanted}
}
\value{
The data with translated data
}
\description{
Select parts of models results from PPBstats::analyse.outputs (eigher parameters comparisons or MCMC).
}
\author{
Gaelle Van Frank
}
\seealso{
\code{\link{PPBstats::analyse.outputs}}, \code{\link{shinemas2R::get.data}}
}

