% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mixture.R
\name{ggplot_mixture1}
\alias{ggplot_mixture1}
\title{Barplot for the comparison of the mixture vs its components & the mean of its components}
\usage{
ggplot_mixture1(res_model, melanges_PPB_mixture, variable,
  plot.type = "comp.in.farm", nb_parameters_per_plot = 8)
}
\arguments{
\item{res_model}{output from the \code{PPBstats::analyse.outputs} function}

\item{variable}{the variable to study}

\item{plot.type}{the type of plot wanted. Can be "comp.in.farm" to compare the mixtures to each of its components in a farm ; "mixVScomp" to compare all mixtures to all components ; "mix.comp.distribution" ; "mix.gain.distribution" to plot the distribution of the difference between mixtures and the mean of its components}

\item{nb_parameters_per_plot}{the number of parameters per plot}

\item{farmers_data}{output from the \code{shinemas2R::get.data} function, with argument query type == "data.mixture.1"}
}
\value{
A list containing, for each environment and mixture, the barplots ("bp") and the tables ("Tab")
}
\description{
Barplot for the comparison of the mixture vs its components & the mean of its components
}
\author{
Gaelle Van Frank
}
\seealso{
\code{\link{shinemas2R::get.data}}
}

