% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_translation.R
\docType{data}
\name{list_translation}
\alias{list_translation}
\title{The translation list}
\format{\preformatted{List of 104
 $ : chr [1:2] "post_winter_global" "note.globale.hiver"
 $ : chr [1:2] "spring_global" "note.globale.printemps"
 $ : chr [1:2] "summer_global" "note.globale.ete"
 $ : chr [1:2] "topography" "topographie"
 $ : chr [1:2] "sowing_practices" "pratiques.semis"
 $ : chr [1:2] "battance" "battance"
 $ : chr [1:2] "micro_field_area" "surface.micro.parcelle"
 $ : chr [1:2] "sowing_qualitative_date" "date.semis"
 $ : chr [1:2] "field_area" "surface.champs"
 $ : chr [1:2] "soil_type_2" "type.de.sol.info2"
 $ : chr [1:2] "soil_type_1" "type.de.sol.info1"
 $ : chr [1:2] "field_info" "info.sur.le.champs"
 $ : chr [1:2] "space_between_micro_field" "espace.entre.micro.parcelles"
 $ : chr [1:2] "sowing_density" "densite.semis"
 $ : chr [1:2] "field_name" "nom.champ"
 $ : chr [1:2] "ploughing_before_sowing" "labour.avant.semis"
 $ : chr [1:2] "drainage" "drainage"
 $ : chr [1:2] "previous_culture" "précédent.cultural"
 $ : chr [1:2] "sowing_special_remarks" "remarques.automne"
 $ : chr [1:2] "sowing_notice_topography_2" "notice.topographie.semis.2"
 $ : chr [1:2] "sowing_notice_sowing_practices" "notice.pratiques.semis"
 $ : chr [1:2] "sowing_notice_battance" "notice.battance"
 $ : chr [1:2] "sowing_notice_micro_field_area" "notice.surface.micro.parcelle"
 $ : chr [1:2] "qualitative_date" "date"
 $ : chr [1:2] "sowing_notice_field_area" "notice.surface.champ"
 $ : chr [1:2] "sowing_notice_soil_type_2" "notice.type.sol.2"
 $ : chr [1:2] "sowing_notice_soil_type_1" "notice.type.sol.1"
 $ : chr [1:2] "sowing_notice_field_info" "notice.info.champ"
 $ : chr [1:2] "sowing_notice_density" "notice.densite.semis"
 $ : chr [1:2] "sowing_notice_field_name" "notice nom.champ"
 $ : chr [1:2] "sowing_notice_ploughing_before_sowing" "notice.labour.avant.semis"
 $ : chr [1:2] "sowing_notice_drainage" "notice.drainage"
 $ : chr [1:2] "sowing_notice_previous_culture" "notice.précédent.cultural"
 $ : chr [1:2] "autumn_rainfall" "pluies.automne"
 $ : chr [1:2] "autumn_temperature" "températures.automne"
 $ : chr [1:2] "autumn_climatic_field_notes" "notes.sur.le.climat.automne"
 $ : chr [1:2] "autumn_climatic_accident" "accidents.climatiques.automne"
 $ : chr [1:2] "post_winter_observation_date" "date.observation.hiver"
 $ : chr [1:2] "post_winter_reprise" "reprise"
 $ : chr [1:2] "post_winter_leaves_attitude" "attitude des feuilles"
 $ : chr [1:2] "post_winter_port_au_tallage" "port.au.tallage"
 $ : chr [1:2] "post_winter_density" "densité"
 $ : chr [1:2] "post_winter_notes" "commentaires.hiver"
 $ : chr [1:2] "post_winter_rainfall" "pluies.hiver"
 $ : chr [1:2] "post_winter_temperature" "températures.hiver"
 $ : chr [1:2] "post_winter_climatic_accident" "accidents.climatiques.hiver"
 $ : chr [1:2] "post_winter_climatic_notes" "notes.sur.le.climat.hiver"
 $ : chr [1:2] "post_winter_field_accident" "accidents.dans.le.champ.hiver"
 $ : chr [1:2] "post_winter_field_notes" "notes.sur.le.champ.hiver"
 $ : chr [1:2] "spring_observation_date" "date.observation.printemps"
 $ : chr [1:2] "spring_tallage" "tallage"
 $ : chr [1:2] "spring_vigueur" "vigueur"
 $ : chr [1:2] "spring_color" "couleur.printemps"
 $ : chr [1:2] "spring_self_propagating" "adventices.printemps"
 $ : chr [1:2] "heading_date_100" "date.épiaison_60-100"
 $ : chr [1:2] "spring_notes_diseases" "notes.maladies"
 $ : chr [1:2] "heading" "remarques.printemps"
 $ : chr [1:2] "spring_heading_note" "note.épiaison"
 $ : chr [1:2] "spring_notes" "commentaires.printemps"
 $ : chr [1:2] "tallage" "tallage"
 $ : chr [1:2] "self_propagating" "adventices"
 $ : chr [1:2] "heading_date_global_a" "date.épiaison.globale.a"
 $ : chr [1:2] "heading_note" "note.épiaison"
 $ : chr [1:2] "spring_rainfall" "pluies.printemps"
 $ : chr [1:2] "spring_temperature" "températures.printemps"
 $ : chr [1:2] "spring_climatic_accident" "accidents.climatiques.printemps"
 $ : chr [1:2] "spring_climatic_notes" "notes.sur.le.climat.printemps"
 $ : chr [1:2] "spring_field_accident" "accidents.dans.le.champ.printemps"
 $ : chr [1:2] "spring_field_notes" "notes.sur.le.champ.printemps"
 $ : chr [1:2] "summer_observation_date" "date.observation.été"
 $ : chr [1:2] "summer_biomass" "biomasse"
 $ : chr [1:2] "heterogeneite" "hétérogénéité"
 $ : chr [1:2] "harvest_date" "date.récolte"
 $ : chr [1:2] "poids_battage" "poids.battage"
 $ : chr [1:2] "rdt_micro_parcelle" "rendement.micro.parcelle"
 $ : chr [1:2] "summer_special_remarks" "commentaires.été"
 $ : chr [1:2] "special_remarks_farmer" "commentaires"
 $ : chr [1:2] "observation_date" "date.observation"
 $ : chr [1:2] "summer_rainfall" "pluies.été"
 $ : chr [1:2] "summer_temperature" "températures.été"
 $ : chr [1:2] "summer_climatic_accident" "accidents.climatiques.été"
 $ : chr [1:2] "summer_climatic_notes" "notes.sur.le.climat.été"
 $ : chr [1:2] "summer_field_accident" "accidents.dans.le.champ.été"
 $ : chr [1:2] "summer_field_notes" "notes.sur.le.champ.été"
 $ : chr [1:2] "temperature" "températures"
 $ : chr [1:2] "rainfall" "pluies"
 $ : chr [1:2] "climate_accident" "accidents.climatiques"
 $ : chr [1:2] "climate_notes" "notes.climatiques"
 $ : chr [1:2] "field_accident" "accidents.dans.le.champ"
 $ : chr [1:2] "tkw" "poids.de.mille.grains"
 $ : chr [1:2] "protein" "taux.de.proteine"
 $ : chr [1:2] "summer_verse" "verse"
 $ : chr [1:2] "plant_height" "hauteur"
 $ : chr [1:2] "spike_weight" "poids.de.l.epi"
 $ : chr [1:2] "awns" "barbe"
 $ : chr [1:2] "curve" "courbure"
 $ : chr [1:2] "color" "couleur"
 $ : chr [1:2] "crosses" "croisements"
 $ : chr [1:2] "disease" "maladie"
  [list output truncated]
}}
\usage{
list_translation
}
\description{
The translation list
}
\author{
Pierre Rivière, Gaelle Van Frank
}
\keyword{datasets}

